add cRecy (#19)
A system based on a DApp that rewards the act of directing the waste towards recycling and composting. The MVP is the backbone of interactions within the system where we started momentum to transfer those interactions to the DApp while it has just started to be developed. It started with actors (recycling and waste generating) submitting waste reports which are the base to minting cRecys and distributing it between actors and the market.