refactor(getTokensInfo): return {tokenId: tokenInfo} instead of grouping by env (#247)
currently we map addresses to tokenInfo objects in the firebase node. As
such, all consumers of the data expect this shape. It might be easier to
refactor the consumers to use multichain data if we use a similar shape
for the new cloud function's output, mapping tokenId to tokenInfo.

also adds "network" field because in the wallet, we'll want to know what
chain a token is on so we can display "Celo Network" or "Ethereum
Network" as a subtitle
([example](https://www.figma.com/file/OE6RFzqXAMqU9ygHtNYyof/Assets-Page-(Multichain)?node-id=2099%3A16352&mode=dev))

I moved supported network ID's to the config and added a code comment on
how we could extend this to allow the client to request whatever network
ID's it wants in the future