refactor(networkId): switch from network to networkId on TokenInfo type (#250)
in the wallet, we will need some mapping from the network field to a
subtitle to show in the transaction feed. if we want to show "Celo
Alfajores Network" in the feed at some point as opposed to just "Celo
Network" (particularly if we enable some "dev mode" in the future),
we'll need information about the environment (testnet vs mainnet) as
well as the network (eth vs celo vs ...). This PR updates TokenInfo to
include that information by using an environment-specific networkId,
rather than just "network".