feat: add script for bulk adding tokens (#324)
This PR adds a script to add new tokens using Coingecko as a data
source. The script:
- gets a list of X tokens by market cap for a given category (i.e.
chain)
- gets the necessary token metadata
- gets and resizes the token image
- saves the updated tokens info json file
- logs warnings if the images are upscaled, stretched, or too big.

I added a new dev dependency to do the image resizing. The resizing gets
us the right dimensions and png format, but of course it doesn't get us
the transparent background etc. I think that will need to be manual.

Related to RET-974