feat: add new tokens for ethereum (#325)
Fixes RET-974

Of the top 100 tokens by marketcap on coingecko for ethereum, the
following tokens could not be added automatically: fantom, astar,
xdce-crowd-sale, zilliqa. These values are the coingecko token id's, but
no token data (i.e. address) could be fetched from their endpoint.

The script in #324 was used to generate these changes - the script has
given the tokens the following default properties:

```
      name,
      symbol,
      decimals,
      address,
      imageUrl: `https://raw.githubusercontent.com/valora-inc/address-metadata/main/assets/tokens/${symbol}.png`,
      isNative: false,
      infoUrl: `https://www.coingecko.com/en/coins/${id}`,
```

These tokens are not eligible for swaps right now, it seems like we
should check the liquidity separately before enabling the ability to
swap into these tokens. Unfortunately the liquidity information is not
directly available in the requests made in #324.